feat: parse custom attributes
Custom attributes are used to tell us some properties of types, functions, parameters and fields that can't be found otherwise in the metadata.

For example, if a function has a  `DocumentationAttribute`, the value will contain the url to the documentation for that function. If a function has a `UnicodeAttribute`, it means that this function uses unicode strings as input, as opposed to ansi strings.