feat: handle parameter indices correctly
In the params table, there are three columns: flags, sequence and name.
For some unknown reason, there are parameters with a sequence of 0.
I have not seen this in the metadata of other clr assemblies.
I'm opting to decrement the param list value so that it starts on index 0.
I can then conditionally increment the index by 1 when printing the params
if the param list points to a list of params that starts with sequence 0 instead of sequence 1.