refactor: use multiple string builders instead of a single plain string for emitting
Up until now, only functions have been emitted. This hasn't required any sophisticated handling of what to emit when. In the next commit we additionally emit the libraries that each function in the namespace resides in. For this, we use separate string builders so that we can build up the list of libraries and list of functions separately. Then in the end we emit the entire contents of each string builder.