chore: emit to .vmodules in home dir instead of out folder
If you use vs code and you want to see the output, I highly recommend adding the .vmodules folder as a workspace folder. Then you have this project in one folder, and the output in a folder below.

When outputting to .vmodules, the v compiler may complain because of duplicate declarations of some functions if you try to compile again after emitting. You therefore might need to delete the output windows folder.